requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
blobfile==2.0.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==2.0.0.post2
filelock==3.12.4
Flask==2.0.1
Flask-WTF==1.1.1
frozenlist==1.4.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
multidict==6.0.4
openai==0.28.1
psycopg2==2.9.6
pycryptodomex==3.19.0
python-dotenv==1.0.0
regex==2023.8.8
requests==2.31.0
tiktoken==0.5.1
tqdm==4.66.1
urllib3==2.0.6
Werkzeug==2.3.6
WTForms==3.0.1
yarl==1.9.2
